repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/1/LLMatic/app.py
from flask import Flask, request

app = Flask(__name__)

# Mock database
users_db = {}
messages_db = {}

@app.route('/user', methods=['POST'])
def create_user():
	data = request.get_json()
	users_db[data['email']] = data
	return {'message': 'User created'}, 201

@app.route('/user/<email>', methods=['GET'])
def get_user(email):
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	return user, 200

@app.route('/signup', methods=['POST'])
def signup():
	data = request.get_json()
	if 'email' not in data or 'password' not in data:
		return {'message': 'Email and password are required'}, 400
	if data['email'] in users_db:
		return {'message': 'Email already registered'}, 400
	users_db[data['email']] = {'password': data['password']}
	return {'message': 'User registered'}, 201

@app.route('/recover_password', methods=['POST'])
def recover_password():
	data = request.get_json()
	if 'email' not in data:
		return {'message': 'Email is required'}, 400
	user = users_db.get(data['email'])
	if not user:
		return {'message': 'User not found'}, 404
	# In a real application, the new password would be emailed to the user
	new_password = 'new_password'
	user['password'] = new_password
	return {'message': 'Password reset', 'new_password': new_password}, 200

@app.route('/user/<email>/profile_picture', methods=['POST'])
def set_profile_picture(email):
	data = request.get_json()
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	user['profile_picture'] = data['profile_picture']
	return {'message': 'Profile picture updated'}, 200

@app.route('/user/<email>/status_message', methods=['POST'])
def set_status_message(email):
	data = request.get_json()
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	user['status_message'] = data['status_message']
	return {'message': 'Status message updated'}, 200

@app.route('/user/<email>/privacy_settings', methods=['POST'])
def set_privacy_settings(email):
	data = request.get_json()
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	user['privacy_settings'] = data['privacy_settings']
	return {'message': 'Privacy settings updated'}, 200

@app.route('/user/<email>/block', methods=['POST'])
def block_contact(email):
	data = request.get_json()
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	user['blocked_contacts'] = user.get('blocked_contacts', [])
	user['blocked_contacts'].append(data['blocked_contact'])
	return {'message': 'Contact blocked'}, 200

@app.route('/user/<email>/unblock', methods=['POST'])
def unblock_contact(email):
	data = request.get_json()
	user = users_db.get(email)
	if not user:
		return {'message': 'User not found'}, 404
	user['blocked_contacts'] = user.get('blocked_contacts', [])
	if data['unblocked_contact'] in user['blocked_contacts']:
		user['blocked_contacts'].remove(data['unblocked_contact'])
	return {'message': 'Contact unblocked'}, 200

@app.route('/group', methods=['POST'])
def create_group():
	data = request.get_json()
	group_id = len(users_db) + 1
	users_db[group_id] = {'name': data['name'], 'picture': data['picture'], 'members': data['members']}
	return {'message': 'Group created', 'group_id': group_id}, 201

@app.route('/group/<int:group_id>', methods=['GET'])
def get_group(group_id):
	group = users_db.get(group_id)
	if not group:
		return {'message': 'Group not found'}, 404
	return group, 200

@app.route('/group/<int:group_id>/add_member', methods=['POST'])
def add_member(group_id):
	data = request.get_json()
	group = users_db.get(group_id)
	if not group:
		return {'message': 'Group not found'}, 404
	group['members'].append(data['member'])
	return {'message': 'Member added'}, 200

@app.route('/group/<int:group_id>/remove_member', methods=['POST'])
def remove_member(group_id):
	data = request.get_json()
	group = users_db.get(group_id)
	if not group:
		return {'message': 'Group not found'}, 404
	if data['member'] in group['members']:
		group['members'].remove(data['member'])
	return {'message': 'Member removed'}, 200

@app.route('/group/<int:group_id>/set_details', methods=['POST'])
def set_group_details(group_id):
	data = request.get_json()
	group = users_db.get(group_id)
	if not group:
		return {'message': 'Group not found'}, 404
	group['name'] = data['name']
	group['picture'] = data['picture']
	return {'message': 'Group details updated'}, 200

@app.route('/message', methods=['POST'])
def send_message():
	data = request.get_json()
	message_id = len(messages_db) + 1
	messages_db[message_id] = {'sender': data['sender'], 'receiver': data['receiver'], 'message': data['message'], 'read': False}
	return {'message': 'Message sent', 'message_id': message_id}, 201

@app.route('/message/<int:message_id>', methods=['GET'])
def get_message(message_id):
	message = messages_db.get(message_id)
	if not message:
		return {'message': 'Message not found'}, 404
	return message, 200

@app.route('/message/<int:message_id>/read', methods=['POST'])
def read_message(message_id):
	message = messages_db.get(message_id)
	if not message:
		return {'message': 'Message not found'}, 404
	message['read'] = True
	return {'message': 'Message marked as read'}, 200

@app.route('/message/<int:message_id>/encrypt', methods=['POST'])
def encrypt_message(message_id):
	message = messages_db.get(message_id)
	if not message:
		return {'message': 'Message not found'}, 404
	# In a real application, the message would be encrypted using a secure method
	message['message'] = 'encrypted_' + message['message']
	return {'message': 'Message encrypted'}, 200
